VkClient/requirements.txt
vk_api
kafka-python==2.0.2